Propagate JavaCompilingDepsTsets as nodes in JavaGlobalCodeInfo
Summary:
This is coming out of S550936. We are seeing huge memory usage for robolectric tests on CI, and the source of the excess memory usage is the TSets that we create here: https://www.internalfb.com/code/fbsource/[7ace1cf1c9ed]/tools/build_defs/android/di_root_module_jar.bzl?lines=59-60 (and to a lesser extent, for Metagen).

We originally did that in D70885383 because for a single build, ensuring TSets for every node all the way up the graph was a bit wasteful (this was actually much more true for metagen where the nodes were often empty anyway, and less true for DI because it is used almost everywhere, so that diff wasn't really a big win the first place, unlike the metagen equivalent D70696221).

The issue when running a lot of robolectric tests (and we have > 600 in the shard I was investigating) is that it means that we end up creating a TSet per target, per node in the graph that uses DI. Given these graphs are often huge, we end up with approx 25 _million_ DICE nodes to represent these TSets (that would suggests around 40k nodes per test using DI, which isn't unreasonably for "core" tests, since our deps are not very modular at all, unfortunately). For comparison, building fb4a uses about 5M DICE nodes.
We end up using about 150M at peak RSS when building, and a lot of that is thanks to the extra DICE nodes (and edges, we have > 200M edges too).

To fix that, we want to avoid recreating copies of all the `JavaCompilingDepsTSet`s in the global code rules. To do that, we need to get the `JavaCompilingDepsTSet` directly in the global code rule. We could either (a) use the `JavaCompilingDepsTSet` that we currently receive, or (b) instead of getting a list of `JavaClasspathEntry` objects when we traverse the TSet in the global code rule, get a list of `JavaCompilingDepsTSet`. We can't do (a) because that was the source of the memory regression that motivated D70696221 and D70885383, so here we do (b). We create a new TSet that is only used for propagating `JavaCompilingDepsTSet` up the graph, and use that in `get_global_code_info`.

Reviewed By: NavidQar

Differential Revision: D80453303

fbshipit-source-id: 85628e15d6143c148ad22176a2e57f9a8765ab2b