Filter non-python libraries from input target deps
Summary:
The current way we get deps in `pyrefly.bxl` means that we must specify the individual targets we want to check, and from that, we can perform the query for its dependencies.

The problem with this is that when performing a wildcard query (`cell//path/to/directory/...`), non-`python_library` targets will be included. This diff fixes that by performing a `kind` filter for `python_library` target types.

Reviewed By: grievejia

Differential Revision: D80489838

fbshipit-source-id: 6a38959ef301063606606727d25caea42938f037