Dedupe buck-headers for cxx rules
Summary:
The buck-headers artefacts are internal to the rules and should in theory be relatively safe to migrate, at least for cxx; however, many scripts rely on the internal layout and properties of buck-headers and may break. I should have updated most if not all of them, but if this still breaks, we may consider rolling out only for {c,cxx}_compile actions.

N.B.: After this, we should be able to merge the buck-headers for compile and precompile, but this diff does not do that (yet).

#buildall

Reviewed By: christycylee

Differential Revision: D80425390

fbshipit-source-id: 9499a22796f1b2712893136a7997d3411066bbf9