Add Layout Render Config
Summary:
Buck2 Rules docs and API docs are currently rendered the same way. This doesn't seem right to me as it forces the layout of both to be the same and always results in one to not be as intuitive.

I'm adding a layout render config sto allow for rendering different docs differently

Reviewed By: JakobDegen

Differential Revision: D80124726

fbshipit-source-id: 4d8281b25f571d78736cfbb09e712ad9eb31b921