Produce js_library providers in all circumstances
Summary:
D40024091 made it so that for Android we skipped creating providers when we're building a configured branch purely for the native code. This was to reduce the cost of analysis on configured branches where we wouldn't even run the actions that the analysis produced.

However, in D80289639, we have a native library that depends upon the JS bundle, so we need to produce the JS bundle for all of our configured branches.

The memory that we were saving before is not so important now, because our main apps use debug builds with a single `arm64` architecture, meaning that this change has no effect on them.

Meta: see https://fb.workplace.com/groups/rn.support/posts/29193730880248816/ for details

Reviewed By: robhogan

Differential Revision: D80691428

fbshipit-source-id: ab0af84071a4aeedc0088205f798fdb5fb6fc77c