Use set of flavors instead of multiple boolean flags
Summary:
This change gives flavors the ability to be combined. So if you want `optimized` but not `pic` now it is possible. Previously it was hardcoded.

Note that here we are still using `optimization_compiler_flags_EXPERIMENTAL` from toolchain, this will be changed to flavor flags soon. So no changes in behaviour is expected, only more extensible code in the future.

Reviewed By: yahayaohinoyi

Differential Revision: D78023439

fbshipit-source-id: 905e71dbb02ce87f8995528e5c37fe63a1b16520