Fixing protobuf bzl for Windows genrules
Summary:
cp does not exist in cmd.exe; it does in powershell,  but figured changing to powershell was a bit too drastic.
forward slash `/` on copy commands target dir was causing the target dir to be read as TMP with a slash argument named whatever the variable ends up being called. Changing that to \\ fixes the error in parsing from copy.exe.

Reviewed By: pkobernik, tschwalensn

Differential Revision: D80802222

fbshipit-source-id: 68134a5589e6309859c17c5ae6afd34194a0cc0f