Back out "D68817857 [split_dwarf] Always create .dwp from .so or executable, not from .dwo directly"
Summary:
With [content-based path hashing](https://fb.workplace.com/groups/2567537826781973/permalink/2713613368841084/) for `cxx_compile`, the compiler will write its output files, including object and DWO files, in a temporary folder. This means that the `DW_AT_dwo_name` DIE inside the object file will point to an invalid location after the output artefacts are renamed.

This means that tools, including `llvm-dwp`, cannot rely on the path embedded in the object files (and propagated into binaries) to find the DWO files. The solution here is to pass the DWO files directly to `llvm-dwp` instead.

Reviewed By: rmaz, nhatminhle

Differential Revision: D80795413

fbshipit-source-id: c83877ead3c62ee6133bb0f419b3b30f19c489e5