Allow the test_cli_lib preamble to be specificed as app env
Summary:
# Context
Currently we rely on buck2 config to override the test-shell config (useful for implementing workflows like code-lenses, etc). The problem is that changing config, for example from the command line, leads to a full invalidation, which requires re-traversing the project, which can add a big overhead when using this feature

# Problem
We would like to pass the preamble as a CLI argument, so it can be done by the caller without relying on config, but at the moment, the preamble is hardwired by the erlang_test rule itself.

# This diff
As a first step, we make it so that the preamble, if given as an app env, is interpreted by the test-shell itself.

Here we still hardwire the preamble in the erlang_test rule, by setting the value of the app environment there, but this will change next

Reviewed By: TD5

Differential Revision: D81128520

fbshipit-source-id: 22d2dba12e7db8da1df3f69515b561e959b09b62