Stop using config for overriding test-shell preamble
Summary:
# Context
Currently we rely on buck2 config to override the test-shell config (useful for implementing workflows like code-lenses, etc).

# Problem
Changing config, for example from the command line, leads to a full invalidation, which requires re-traversing the project, which can add a big overhead when using this feature.

# This diff
In the previous diff we set things up so that an application env can be used to set the preamble (which a coller can pass a command-line arg).

Here we remove the attr that was being used for this, and no longer pass the preamble from the erlang_test rule.

Reviewed By: TD5, TheGeorge

Differential Revision: D81128519

fbshipit-source-id: 6b02a79d2136e72e6ff072a08f48a86908364fd3