Fix showcase header issue
Summary:
Adds `apple_static_archive` to wrap the binaries of showcaseable targets.

I think I advised originally that it was unnecessary, because the main feature of `apple_static_archive` is to control which dependencies are included in a (Xc)framework definition, and we don't want any dependencies, which is the default thing you get with `apple_library`.

However, they also offer some control over which headers get included - as it ensures only the headers from the actually included binaries are added to the Xcframework. Without this, ObjC targets with ObjC dependencies will do the wrong thing (in the Showcase context, at least).

It's quite possible the current logic for assembling the headers in an Xcframework is not quite right. I seem to recall there's really no 'good' default answer, but it might be worth investigating again.

Differential Revision: D81233693

fbshipit-source-id: 84abea92dbdeb77955459e82038413496bdb15c0