Add flag to native planner to dump linker invocation
Summary:
For distributed thin-lto thin-link, the linker is invoked through two layers of wrappers, some of which modify the arguments. This makes it difficult to debug the linker invocation. This diff adds a flag to the outermost wrapper, a C++ program, that makes the wrapper dump the linker invocation instead of actually running it.

This makes it easier to debug thin-link failures.

Reviewed By: kyulee-com

Differential Revision: D81166353

fbshipit-source-id: 43692240c809215874c52dc03fefd04a13968247