Make header units compile with normal #include lines
Summary:
Instead of using the cli -include flag to precompile
headers, which causes clang to mostly give useless paths in
diagnostics, write and use normal #include lines in the previously-empty
stub header. This isn't writing any more files to disk (maybe less even)
because we're putting what was previously in an argfile into a
file that existed but was empty.

Reviewed By: nhatminhle

Differential Revision: D80378798

fbshipit-source-id: 6e7714e66554afa14b47cd0848ba3a6a8ab87d1f