vsgo: Emit toolchain compiler settings into generated vcxprojs.
Summary:
Previously, if you set compiler flags and/or preprocessor definitions in your toolchain (e.g. via `CxxCompilerInfo`) these would not make it into the generated vcxproj's, leading to IntelliSense squigglies for system headers and certain other cases.

Making this work is not trivial because the way that vsgo currently operates is by inspecting attributes of rules, which doesn't work in this case because the values we need are all stuffed away into `cmd_args` inside of a provider.  To solve this, for toolchains we have to run analysis on the toolchain target  to get the `CxxCompilerInfo`, then run a dynamic action that will resolve the `cmd_args` into actual paths and flags, and materialize a file with these values.  Then read these back later in the continuation to parse them out.

Incidentally, a similar strategy would also work to solve a current limitation of vsgo that the file extension is hardcoded to `.o`, even though you can change this in your toolchain (and a more natural extension in windows is `.obj`).  This was discussed in https://github.com/facebook/buck2/issues/868 but we did not arrive at a solution at the time.  Now, we could simply write out the extension from the provider to the same settings file and read it back later.  That isn't handled in this PR, but it's a nice consequence of this PR that the groundwork is laid for more advanced functionality such as the aforementioned in the future.

X-link: https://github.com/facebook/buck2/pull/1074

Reviewed By: autozimu

Differential Revision: D80191427

Pulled By: scottcao

fbshipit-source-id: 047a1aa1a60cb88450d403fd760ace534d207707