Use dep_files in "check" subtarget to track which headers used
Summary:
Reference: https://buck2.build/docs/rule_authors/dep_files/

The dep file for the main compile action is set here:

https://www.internalfb.com/code/fbsource/[d60258ce4078468ce4e9ad6b7e6cc1cd7354955f]/fbcode/buck2/prelude/cxx/compile.bzl?lines=342%2C349-357%2C442-443%2C446%2C451

This diff applies the same logic to the typecheck-only compile action.

This makes it visible to Buck, out of all the headers available for this C++ compilation to read, which ones actually are read and which ones are unused. If some header has changes during a subsequent build, Buck can now bypass re-checking source files in which that header is unused.

Reviewed By: IanChilds

Differential Revision: D81468244

fbshipit-source-id: 29b38dc30962edf0cbfd6ccfb5fd6cae2b23d45c