erlang: simplify generated sources handling
Summary:
We weren't using the map result - a simple list is enough, we can also simplify the code by moving where the output path is generated (and how we generate it now that we don't have multiple toolchains).

The rejection of modules from `src` was a noop, unless there was a duplicate module - one generated and one in sources, but we should absolutely error if that is the case, rather than silently ignoring it as before.

Reviewed By: TheGeorge

Differential Revision: D81781359

fbshipit-source-id: e57765d723e40546c862697e75b4859f407d1c7b