Add `[resources]` subtarget to `apple_resource_bundle()`
Summary:
**Context**

`apple_resource_bundle()` is a special rule that has no outputs and exists to simulate execution groups.

**Problem**

There's a need to inspect the outputs/resources produced for debugging purposes.

**Solution**

Expose via a `[resources]` subtarget. Do not expose as a default output, as to keep the notion that this is a special target without output.

Reviewed By: IanChilds

Differential Revision: D82015182

fbshipit-source-id: aabf811a5be2b209d4644475592f8c5c3514daec