Swift: add `[swift.check]` to `apple_library()`
Summary:
Introduces a `[swift.check]` subtarget which typechecks any Swift code in that `apple_library()`.

Note that `[check]` subtarget is reserved for non-Swift code and we don't want to change that behaviour (nor do we want to just add the Swift checks to the pre-existing target) (for behaviour preservation purposes).

Reviewed By: rmaz

Differential Revision: D82103171

fbshipit-source-id: 2473559e66024c9a996b98c9a2ef18df381d283a