Add ability to query classpath of apk_genrule
Summary:
Unfortunately, this not being present on `apk_genrule` means that if you have a target that does `query(classpath(:my_apk_genrule))`, the query returns nothing instead of the classpath. That means that in turn any audit dependencies test on an `apk_genrule` is currently useless (see https://fb.workplace.com/groups/askbuck/posts/29404591159162786/).

So, it's pretty likely this is going to break a bunch of tests. Those breakages are CORRECT, and you should not revert this diff. You should either fix the test, or if you don't care about it, then delete the test.

Reviewed By: a-coox

Differential Revision: D82207205

fbshipit-source-id: 212caa40ac462bebd1db5fcf411b4bf25b974a01