Export `get_exported_preprocessor_args` function from `preprocessor.bzl` so it can be reused.
Summary:
Internally, we have a rule that generates cxx headers from some internal idl-style language.  Right now we do some hackery to make a `cxx_library` that passes `exported_preprocessor_flags=["-I$(location :gen-target)"]` but this requires some hoops to jump through, and we'd like to simply return a `CPreprocessorInfo` so that cxx_library targets can simply take a dep.  We have this working internally but it requires calling this function, which is prefixed with an `_` so cannot be imported.  Just make this function public.

X-link: https://github.com/facebook/buck2/pull/1098

Reviewed By: JakobDegen

Differential Revision: D82260585

fbshipit-source-id: a61bec9072186bca47c7577cb3ae3b137e3af719