Make branches for raw and compressed more similar in SecondaryDexCompressionExecutableMain
Summary:
This diff extracts the expression `rawSecondaryDexesDirPath.resolve(secondaryDexName)` into a variable `secondaryDexName`. That is exactly what we do on the branch `compression != "raw"`:

https://www.internalfb.com/code/fbsource/[3db1aa79035c0773fb1217575fe01d216373a311]/fbcode/buck2/prelude/toolchains/android/src/com/facebook/buck/android/dex/SecondaryDexCompressionExecutableMain.java?lines=113%2C118%2C122%2C127%2C130-131%2C162

This makes the code a bit easier to reason about since the two branches are now more similar (and might allow us to unify them later on). The behavior is unchanged.

Reviewed By: IanChilds

Differential Revision: D82406144

fbshipit-source-id: 5cf77f2e33c4bd113dee2d643ba308cd24c381ae