Add option to skip generating base and analysis platform
Summary:
# Problem

It's currently not possible to call `apple_target_platforms` twice with the same base name (eg. for different cxx platforms like separating iOS from MacCatalyst) because it registers a platform with `<base_name>` and `<base_name>-analysis`.
```
error: Attempted to register target waios//platforms:consumer twice, re-run the command with `--stack` to obtain a call stack of the first registration
       --> fbcode/buck2/prelude/platforms/apple/platforms.bzl:131:9
        |
    131 |           platform_rule(
        |  _________^
    132 | |             name = base_name,
    133 | |             constraint_values = constraint_values,
    134 | |             visibility = visibility,
    135 | |             deps = deps,
    136 | |         )
        | |_________^
        |
```

# Solution
Add an option to skip generating the `<base_name>` and `<base_name>-analysis` platforms. This option is defaulted to preserve the existing functionality.

Reviewed By: rmaz

Differential Revision: D82446918

fbshipit-source-id: 8159843065082efd24afe54146ab59bcf7b6d08d