Enable dwo path rewriting for fbcode split debug
Summary:
Turn on the rewriting feature implemented in D82482681, passing it the list of debug info files as part of the link action.

This diff enables the rewriting whenever content-based paths are supported (hence might be present), although they are not turned on until D81327920. This exercises the no-op rewrite code path ensuring that if there is nothing to rewrite, the script does not alter debug infos.

Reviewed By: christycylee

Differential Revision: D82482680

fbshipit-source-id: fd7efe986b651ef591b81383864a6ff8e7c8a811