Allow transformation spec to be applied to first order deps
Summary:
This will be used for enabling focused debugging for first order deps.
We need to use several matchers for transformation specs. And "debug" transformations needed to be determined based on graph state. Because we want to enable debugging for root target and it's first order deps. This change introduces matchers for "first order deps" of a target target.
```
transformation_spec(
    name = "debug_a_target",
    transformations = [
        ("graph:first_order_deps", "debug"),
    ],
)
```

Differential Revision: D83142713

fbshipit-source-id: e9886c5374bd4b532c19f085eca1278e2896a422