Add pattern matching in transformation spec
Summary:
Re-using API we have in link groups that allows us to match whole folders

We need to use several matchers for transformation specs. The first one was implemented as an exact match
```
transformation_spec(
    name = "optimize_a_target",
    transformations = [
        ("//some/folder:concrete_target", "optimized"),
    ],
)
```
and mostly useful for testing.
In real world we want to mach by whole folders
```
transformation_spec(
    name = "optimize_a_target",
    transformations = [
        ("//some/folder/...", "optimized"),
    ],
)
```
and enable optimisation for a cohort of target.

Differential Revision: D83142710

fbshipit-source-id: 26fcf2272501f442c5ebbe7c3b8fa60e8cb44243