Created build graph pattern matching
Summary:
We need to use several matchers for transformation specs. And "debug" transformations needed to be determined based on graph state. Because we want to enable debugging for root target and it's first order deps. This change introduces matchers for "root" target that will return match for the source code of `cpp_executable` or `cpp_unittest` itself.
```
transformation_spec(
    name = "debug_a_target",
    transformations = [
        ("graph:root", "debug"),
    ],
)
```

Reviewed By: yahayaohinoyi

Differential Revision: D83142709

fbshipit-source-id: 291dec39d98baf0f91a3984c2918ff0f9038a6f8